Add user & room ID to info command
So, these were something I was hunting around for, and thought could be
useful as basic information to pass along, especially when getting
started with a new chat system.

I'm not in love with that blank room ID in the test specs, I'll see if I
can add something to fill that in.