Trigger :message_received whenever robot.receive is called
Certain handlers use the chat route regex `/./` to permissively trigger
a method and evaluate some sort of conditional logic therein. This
effectively means that `:unhandled_message` is never triggered, which
breaks functionality like invalid command logging in lita-metrics. With
this change, plugin authors will no longer have to use the permissive
regex as a workaround. Please see the following for an example:
https://github.com/tristaneuan/lita-google-translate/pull/1